requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
cymem==2.0.11
datasketch==1.6.5
decorator==5.2.1
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.5.0/en_core_web_trf-3.5.0-py3-none-any.whl#sha256=8902305d2ced83d98a8e88efd93ea8970a70dda9bb24b0024a8b798d1cc913d5
exceptiongroup==1.3.0
executing==2.2.0
feedparser==6.0.11
filelock==3.13.1
frozenlist==1.6.0
fsspec==2024.6.1
huggingface-hub==0.31.2
idna==3.10
ipython==8.36.0
jedi==0.19.2
Jinja2==3.1.4
jsonpickle==4.0.5
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.4.3
murmurhash==1.0.12
networkx==3.3
numpy==2.2.5
packaging==25.0
pandas==2.2.3
parso==0.8.4
pathlib_abc==0.1.1
pathy==0.11.0
pillow==11.0.0
preshed==3.0.9
prompt_toolkit==3.0.51
propcache==0.3.1
pure_eval==0.2.3
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-louvain==0.16
pytz==2025.2
pyvis==0.3.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scipy==1.15.3
sgmllib3k==1.0.0
six==1.17.0
smart-open==6.4.0
spacy==3.8.4
spacy-alignments==0.9.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.2.5
srsly==2.5.1
stack-data==0.6.3
sympy==1.13.3
tenacity==9.1.2
thinc==8.3.6
tokenizers==0.13.3
torch==2.7.0+cu118
torchaudio==2.7.0+cu118
torchvision==0.22.0+cu118
tqdm==4.67.1
traitlets==5.14.3
transformers==4.30.2
typer==0.9.4
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
yarl==1.20.0
